refactor: Move exceptions, fix linters
