+++ botx/bot/api/responses/command_accepted.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict
 
 
-def build_accepted_response() -> Dict[str, Any]:
+def build_command_accepted_response() -> Dict[str, Any]:
     """Build accepted response for BotX.
 
     It should be sent if the bot started processing a command.
