+++ botx/bot/middlewares/exception_middleware.py
