+++ botx/client/chats_api/create_chat.py
@@ -3,10 +3,7 @@ from uuid import UUID
 
 from botx.client.authorized_botx_method import AuthorizedBotXMethod
 from botx.client.botx_method import response_exception_thrower
-from botx.client.chats_api.exceptions import (
-    ChatCreationError,
-    ChatCreationProhibitedError,
-)
+from botx.client.exceptions.chats import ChatCreationError, ChatCreationProhibitedError
 from botx.models.api_base import UnverifiedPayloadBaseModel, VerifiedPayloadBaseModel
 from botx.models.enums import APIChatTypes, ChatTypes, convert_chat_type_from_domain
