+++ botx/client/exceptions/chats.py
