+++ botx/client/exceptions/files.py
