+++ botx/client/exceptions/users.py
