+++ botx/client/files_api/download_file.py
@@ -7,8 +7,8 @@ from botx.async_buffer import AsyncBufferWritable
 from botx.client.authorized_botx_method import AuthorizedBotXMethod
 from botx.client.botx_method import response_exception_thrower
 from botx.client.exceptions.common import ChatNotFoundError
+from botx.client.exceptions.files import FileDeletedError, FileMetadataNotFound
 from botx.client.exceptions.http import InvalidBotXStatusCodeError
-from botx.client.files_api.exceptions import FileDeletedError, FileMetadataNotFound
 from botx.models.api_base import UnverifiedPayloadBaseModel
