+++ botx/client/users_api/search_user_by_login.py
@@ -1,6 +1,6 @@
 from botx.client.authorized_botx_method import AuthorizedBotXMethod
 from botx.client.botx_method import response_exception_thrower
-from botx.client.users_api.exceptions import UserNotFoundError
+from botx.client.exceptions.users import UserNotFoundError
 from botx.client.users_api.user_from_search import BotXAPISearchUserResponsePayload
 from botx.models.api_base import UnverifiedPayloadBaseModel
