+++ setup.cfg
@@ -46,32 +46,32 @@ nested_classes_whitelist = Config
 allowed_domain_names = data, handler, result, content, file
 
 per-file-ignores =
-    botx/bot/bot.py:WPS203
+    botx/bot/bot.py:WPS203,
     botx/__init__.py:WPS203,WPS410,WPS412,F401,
-    botx/bot/api/entities.py:WPS202,
-    botx/bot/api/enums.py:WPS202,
     # https://github.com/wemake-services/wemake-python-styleguide/issues/2172
-    botx/bot/handler_collector.py:WPS437
-    botx/client/notifications_api/internal_bot_notification.py:WPS202
-    botx/client/notifications_api/mentions.py:WPS202,WPS226
+    botx/bot/handler_collector.py:WPS437,
+    botx/client/notifications_api/internal_bot_notification.py:WPS202,
+    # Complex model converting
+    botx/models/message/incoming_message.py:WPS232,
+    # WPS reacts at using `}` in f-strings
+    botx/models/message/mentions.py:WPS226,
+
     tests/*:E501,WPS110,WPS114,WPS116,WPS118,WPS202,WPS221,WPS226,WPS237,WPS402,WPS420,WPS428,WPS430,WPS432,WPS441,WPS442,WPS520,S105,S106
-    botx/shared_models/domain/attachments.py:WPS202
-    botx/bot/api/attachments.py:WPS202,WPS212
-    # TODO: Remove ignore after improving models structure
-    botx/client/notifications_api/markup.py:WPS202
 
 ignore =
     # TODO: Drop this ignores on writing docs
-    DAR101, DAR201, D202, DAR401, RST202, RST215
+    DAR101, DAR201, D202, DAR401, RST202, RST215,
 
-    # Class without base class
-    WPS306,
     # Upper-case constant in class
     WPS115,
+    # Too many module members
+    WPS202,
     # Too many arguments
     WPS211,
     # f-strings
     WPS305,
+    # Class without base class
+    WPS306,
     # Implicit string concatenation
     WPS326,
     # Explicit string concatenation
